requirements.txt
asgiref==3.7.2
certifi==2023.11.17
Django==3.2.23
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
wincertstore==0.2
langchain==0.0.352
openai==1.6.0
chromadb==0.4.21
tiktoken==0.5.2
mysqlclient==2.2.1
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.4
python-dotenv==1.0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.21
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
Django==3.2.23
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
exceptiongroup==1.2.0
fastapi==0.103.2
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
# joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1691577114857/work
jsonpatch==1.33
jsonpointer==2.4
kubernetes==28.1.0
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
marshmallow==3.20.1
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.1
# numpy @ file:///D:/bld/numpy_1700874645773/work/dist/numpy-1.26.2-cp310-cp310-win_amd64.whl#sha256=8cfb1ae9bf54a72d05ed128f65e82ba74a830d2e1ca1cb9956b2a805e79a383e
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.6.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
posthog==3.1.0
protobuf==4.25.1
pulsar-client==3.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
PyPika==0.48.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scikit-surprise==1.1.3
# SciPy @ file:///C:/bld/scipy-split_1700812632017/work/dist/scipy-1.11.4-cp310-cp310-win_amd64.whl#sha256=9160b072e08f92f87a3e6fce6e98a464aa6c326f2845139fa1f26af41c3c8580
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.24
sqlparse==0.4.4
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.25.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wincertstore==0.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0